Safer debugging ASSERT() macro.
Spl's ASSERT macro uses the typical do/while(0) form for if/else
friendliness, however, this limits its use in contexts where a a do loop
is not valid; such as within another multi-statement style macro.

This commit converts the branch-predicition logic in ASSERT() to use
the && and also converts it to use the new PANIC1() function.  PANIC1()
is a single-string argument inline function version of PANIC(), suitable
for use within ASSERT().